:bug: Refactor DesktopDeviceManager to DeviceManager for multi-platform support
