+++ composeApp/src/commonMain/kotlin/com/crosspaste/sync/DeviceListener.kt
@@ -1,0 +1,10 @@
+package com.crosspaste.sync
+
+import com.crosspaste.dto.sync.SyncInfo
+
+interface DeviceListener {
+
+    fun addDevice(syncInfo: SyncInfo)
+
+    fun removeDevice(syncInfo: SyncInfo)
+}
