+++ composeApp/src/commonMain/kotlin/com/crosspaste/sync/DeviceManager.kt
@@ -1,12 +1,83 @@
 package com.crosspaste.sync
 
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateListOf
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.setValue
+import com.crosspaste.app.AppInfo
+import com.crosspaste.config.ConfigManager
+import com.crosspaste.dao.sync.SyncRuntimeInfoDao
+import com.crosspaste.dao.sync.createSyncRuntimeInfo
 import com.crosspaste.dto.sync.SyncInfo
+import com.crosspaste.utils.getJsonUtils
+import io.github.oshai.kotlinlogging.KotlinLogging
 
-interface DeviceManager {
+class DeviceManager(
+    private val appInfo: AppInfo,
+    private val configManager: ConfigManager,
+    private val syncManager: SyncManager,
+    private val syncRuntimeInfoDao: SyncRuntimeInfoDao,
+) : DeviceListener {
 
-    val searching: Boolean
+    private val logger = KotlinLogging.logger {}
 
-    val syncInfos: MutableList<SyncInfo>
+    private val jsonUtils = getJsonUtils()
 
-    fun refresh()
+    var searching by mutableStateOf(false)
+
+    private val allSyncInfos: MutableMap<String, SyncInfo> = mutableMapOf()
+
+    var syncInfos: MutableList<SyncInfo> = mutableStateListOf()
+
+    private val isSelf: (String) -> Boolean = { appInstanceId ->
+        appInstanceId == appInfo.appInstanceId
+    }
+
+    private val isNew: (String) -> Boolean = { appInstanceId ->
+        !syncManager.getSyncHandlers().keys.contains(appInstanceId)
+    }
+
+    private val isBlackListed: (String) -> Boolean = { appInstanceId ->
+        val blackSyncInfos: List<SyncInfo> =
+            jsonUtils.JSON.decodeFromString(
+                configManager.config.blacklist,
+            )
+        blackSyncInfos.map { it.appInfo.appInstanceId }
+            .contains(appInstanceId)
+    }
+
+    fun refresh() {
+        searching = true
+        try {
+            syncInfos.clear()
+            val existSyncRuntimeInfos =
+                allSyncInfos.filter { !isNew(it.key) }
+                    .map { createSyncRuntimeInfo(it.value) }
+            if (existSyncRuntimeInfos.isNotEmpty()) {
+                syncRuntimeInfoDao.update(existSyncRuntimeInfos)
+                    .forEach {
+                        syncManager.resolveSync(it)
+                    }
+            }
+            syncInfos.addAll(allSyncInfos.filter { isNew(it.key) && !isBlackListed(it.key) }.values)
+        } finally {
+            searching = false
+        }
+    }
+
+    override fun addDevice(syncInfo: SyncInfo) {
+        val appInstanceId = syncInfo.appInfo.appInstanceId
+        logger.debug { "Service resolved: $syncInfo" }
+        if (!isSelf(appInstanceId)) {
+            allSyncInfos[appInstanceId] = syncInfo
+            refresh()
+        }
+    }
+
+    override fun removeDevice(syncInfo: SyncInfo) {
+        val appInstanceId = syncInfo.appInfo.appInstanceId
+        logger.debug { "Service removed: $syncInfo" }
+        allSyncInfos.remove(appInstanceId)
+        refresh()
+    }
 }
